# zoo/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/enclosure": {
            "get": {
                "description": "Возвращает вольер по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enclosures"
                ],
                "summary": "Получение вольера по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID вольера",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.GetAllEnclosuresResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Вольер не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет новый вольер",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enclosures"
                ],
                "summary": "Создание нового вольера",
                "parameters": [
                    {
                        "description": "Данные нового вольера",
                        "name": "enclosure",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.AddEnclosureRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID нового вольера",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет вольер по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enclosures"
                ],
                "summary": "Удаление вольера",
                "parameters": [
                    {
                        "description": "ID вольера",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.DeleteEnclosureRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Вольер удален"
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/enclosures": {
            "get": {
                "description": "Возвращает список всех вольеров",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "enclosures"
                ],
                "summary": "Получение всех вольеров",
                "responses": {
                    "200": {
                        "description": "Список вольеров",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/enclosure.Enclosure"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/feeding": {
            "post": {
                "description": "Кормит животное по расписанию",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeding"
                ],
                "summary": "Кормление животного",
                "parameters": [
                    {
                        "description": "ID животного",
                        "name": "animal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.FeedAnimalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Животное покормлено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Животное не найдено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Животное не голодно",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/feeding/schedule": {
            "post": {
                "description": "Добавляет новое расписание кормления для животного",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeding"
                ],
                "summary": "Создание нового расписания кормления",
                "parameters": [
                    {
                        "description": "Новое расписание кормления",
                        "name": "schedule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.AddFeedingScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ID нового расписания",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/feeding/schedule/interval": {
            "put": {
                "description": "Изменяет интервал кормления для расписания",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeding"
                ],
                "summary": "Изменение интервала кормления",
                "parameters": [
                    {
                        "description": "Изменение интервала кормления",
                        "name": "schedule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.ChangeFeedingIntervalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Интервал кормления изменен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "animal.Type": {
            "type": "string",
            "enum": [
                "predator",
                "herbivore",
                "bird",
                "aquarium"
            ],
            "x-enum-varnames": [
                "Predator",
                "Herbivore",
                "Bird",
                "Aquarium"
            ]
        },
        "enclosure.Enclosure": {
            "type": "object",
            "properties": {
                "animalCount": {
                    "type": "integer"
                },
                "animalIDs": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "object"
                    }
                },
                "animalType": {
                    "$ref": "#/definitions/animal.Type"
                },
                "id": {
                    "type": "integer"
                },
                "maxCapacity": {
                    "type": "integer"
                }
            }
        },
...
        "http.AddEnclosureRequest": {
            "type": "object",
            "properties": {
                "animal_type": {
                    "description": "Тип животного\nrequired: true\nenum: predator,herbivore,bird,aquarium\nexample: aquarium",
                    "allOf": [
                        {
                            "$ref": "#/definitions/animal.Type"
                        }
                    ]
                },
                "capacity": {
                    "description": "Вместимость вольера\nrequired: true\nminimum: 1\nexample: 10",
                    "type": "integer"
                }
            }
        },
        "http.AddFeedingScheduleRequest": {
            "type": "object",
            "properties": {
                "animal_id": {
                    "type": "integer"
                },
                "feeding_interval": {
                    "type": "integer"
                },
                "food_type": {
                    "type": "string"
                }
            }
        },
...
        "http.ChangeFeedingIntervalRequest": {
            "type": "object",
            "properties": {
                "feeding_interval": {
                    "type": "integer"
                },
                "schedule_id": {
                    "type": "integer"
                }
            }
        },
...
        "http.DeleteEnclosureRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID вольера\nrequired: true\nexample: 1",
                    "type": "integer"
                }
            }
        },
        "http.FeedAnimalRequest": {
            "type": "object",
            "properties": {
                "animal_id": {
                    "type": "integer"
                },
                "food": {
                    "type": "string"
                }
            }
        },
...
        "http.GetAllEnclosuresResponse": {
            "type": "object",
            "properties": {
                "animal_count": {
                    "type": "integer"
                },
                "animal_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "animal_type": {
                    "$ref": "#/definitions/animal.Type"
                },
                "id": {
                    "type": "integer"
                },
                "max_capacity": {
                    "type": "integer"
                }
            }
        },
...
    }
}